# docs/swagger.yaml
basePath: /
definitions:
...
  app.ListRes:
    properties:
      list:
        description: Data list // 数据清单
      pager:
        allOf:
        - $ref: '#/definitions/app.Pager'
        description: Pagination info // 翻页信息
    type: object
  app.Pager:
    properties:
      page:
        description: Page number // 页码
        type: integer
      pageSize:
        description: Page size // 每页数量
        type: integer
      totalRows:
        description: Total rows // 总行数
        type: integer
    type: object
  app.Res:
    properties:
      code:
        type: integer
      context: {}
      data: {}
      details: {}
      message: {}
      pageIndex:
        description: '1-BASED wire value: 0/absent = non-paginated message, n>0 =
          download page n-1 (internal 0-based). 1-based so page 0 stays distinguishable
          from non-paginated messages under proto3 zero-value elision; only set for
          paginated download messages on pv>=2 connections. Client-to-server PageAck.pageIndex
          stays 0-based. // 线上值为 1-BASED：0/缺省 = 非分页消息，n>0 = 下载页 n-1（内部 0-based）。采用
          1-based 是为了让第 0 页在 proto3 零值不编码规则下仍可与非分页消息区分；仅 pv>=2 连接的分页下载消息会设置。客户端→服务端的
          PageAck.pageIndex 保持 0-based'
      path: {}
      status:
        type: boolean
      vault: {}
    type: object
...
  dto.FileDTO:
    properties:
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      entryId:
        description: v3 entry UUID (fs_entry.id) // v3 条目 UUID
        type: string
      id:
        description: File ID // 文件 ID
        type: integer
      lastTime:
        description: Updated timestamp // 更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: File path // 文件路径
        type: string
      pathHash:
        description: Path hash // 路径哈希
        type: string
      rename:
        description: Rename flag // 重命名标记
        type: integer
      size:
        description: File size // 文件大小
        type: integer
      updatedAt:
        description: Updated at time // 更新时间
        type: string
    type: object
...
  dto.NoteDTO:
    properties:
      clientName:
        description: Client name // 客户端名称
        type: string
      clientType:
        description: Client type // 客户端类型
        type: string
      clientVersion:
        description: Client version // 客户端版本
        type: string
      content:
        description: Note content // 笔记内容
        type: string
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      entryId:
        description: v3 entry UUID (fs_entry.id) // v3 条目 UUID
        type: string
      id:
        description: Note ID // 笔记 ID
        type: integer
      lastTime:
        description: Record update timestamp // 记录更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: Note path // 笔记路径
        type: string
      pathHash:
        description: Path hash // 路径哈希
        type: string
      size:
        description: Note size // 笔记大小
        type: integer
      updatedAt:
        description: Updated at time // 更新时间
        type: string
      version:
        description: Version number // 版本号
        type: integer
    type: object
...
  dto.NoteModifyOrCreateRequest:
    properties:
      baseHash:
        description: Base hash for sync // 同步基准哈希
        example: bhash789
        type: string
      baseHashMissing:
        description: Marks if baseHash is unavailable // 标记基准哈希是否缺失
        example: false
        type: boolean
      content:
        description: Note content // 笔记内容
        example: '# Hello World'
        type: string
      contentHash:
        description: Content hash // 内容哈希
        example: chash012
        type: string
      context:
        description: Context // 同步上下文
        example: ctx123
        type: string
      createOnly:
        description: If true, fail if note already exists // 如果为 true，笔记已存在则失败
        example: false
        type: boolean
      ctime:
        description: Creation timestamp // 创建时间戳
        example: 1700000000
        type: integer
      isConflictResolved:
        description: Marks if conflict is resolved manually // 标记是否为手动解决冲突
        example: false
        type: boolean
      mtime:
        description: Modification timestamp // 修改时间戳
        example: 1700000000
        type: integer
      path:
        description: Note path // 笔记路径
        example: ReadMe.md
        type: string
      pathHash:
        description: Path hash // 路径哈希
        example: hash123
        type: string
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - path
    - vault
    type: object
...
  dto.NoteWithFileLinksResponse:
    properties:
      content:
        description: Note content // 笔记内容
        type: string
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      fileLinks:
        additionalProperties:
          type: string
        description: Map of file link to actual path // 文件链接到实际路径的映射
        type: object
      lastTime:
        description: Record update timestamp // 记录更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: Note path // 笔记路径
        type: string
      pathHash:
        description: Path hash // 路径哈希
        type: string
      updatedAt:
        description: Updated at time // 更新时间
      version:
        description: Version number // 版本号
        type: integer
    type: object
...
  dto.ShareCancelRequest:
    properties:
      id:
        description: Share ID (optional) // 分享 ID (可选)
        example: 1
        type: integer
      path:
        description: Resource path (optional) // 资源路径 (可选)
        example: ReadMe.md
        type: string
      pathHash:
        description: Resource path Hash (optional) // 资源路径哈希 (可选)
        example: hash123
        type: string
      vault:
        description: Vault name // 保险库名称
        example: defaultVault
        type: string
    required:
    - vault
    type: object
...
host: localhost:9000
info:
  contact:
    email: haierkeys@gmail.com
    name: Haierkeys
    url: https://github.com/haierkeys
  description: This is the Fast Note Sync Service HTTP API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fast Note Sync Service HTTP API
  version: "1.0"
paths:
...
  /api/file:
    delete:
      description: Permanently delete a specific attachment record and its physical
        file
      parameters:
      - description: Context // 同步上下文
        example: ctx123
        in: query
        name: context
        type: string
      - description: File path // 文件路径
        example: Image.png
        in: query
        name: path
        required: true
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        required: true
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Delete attachment
      tags:
      - File
    get:
      description: Get raw binary data of an attachment by path, supports strong cache
        control
      parameters:
      - description: Context // 同步上下文
        example: ctx123
        in: query
        name: context
        type: string
      - description: Is in recycle bin // 是否在回收站
        example: false
        in: query
        name: isRecycle
        type: boolean
      - description: File path // 文件路径
        example: Image.png
        in: query
        name: path
        required: true
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Success
          schema:
            type: file
      security:
      - UserAuthToken: []
      summary: Get attachment content
      tags:
      - File
    post:
      consumes:
      - multipart/form-data
      description: Upload a file as an attachment
      parameters:
      - description: Vault name
        in: formData
        name: vault
        required: true
        type: string
      - description: File path
        in: formData
        name: path
        required: true
        type: string
      - description: Creation timestamp
        format: int64
        in: formData
        name: ctime
        type: integer
      - description: Modification timestamp
        format: int64
        in: formData
        name: mtime
        type: integer
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Upload attachment
      tags:
      - File
...
  /api/note:
    delete:
      description: Move note to trash
      parameters:
      - description: Context // 同步上下文
        example: ctx123
        in: query
        name: context
        type: string
      - description: Note path // 笔记路径
        example: ReadMe.md
        in: query
        name: path
        required: true
        type: string
      - description: Path hash // 路径哈希
        example: hash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Delete note
      tags:
      - Note
    get:
      description: Get specific note content and metadata by path or path hash
      parameters:
      - description: Is in recycle bin // 是否在回收站
        example: false
        in: query
        name: isRecycle
        type: boolean
      - description: Note path // 笔记路径
        example: ReadMe.md
        in: query
        name: path
        required: true
        type: string
      - description: Path hash // 路径哈希
        example: hash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteWithFileLinksResponse'
              type: object
      security:
      - UserAuthToken: []
      summary: Get note details
      tags:
      - Note
    post:
      consumes:
      - application/json
      description: Handle note creation, modification, or renaming (identified by
        path change)
      parameters:
      - description: Note Content
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.NoteModifyOrCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Create or update note
      tags:
      - Note
...
  /api/notes/share-paths:
    get:
      parameters:
      - description: Vault name
        in: query
        name: vault
        required: true
        type: string
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - UserAuthToken: []
      summary: Get active shared note paths
      tags:
      - Share
...
  /api/share:
    delete:
      consumes:
      - application/json
      description: Cancel a share by ID or path parameters
      parameters:
      - description: Cancel Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.ShareCancelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Cancel share
      tags:
      - Share
    get:
      description: Get share token and info by vault and path
      parameters:
      - description: Resource path // 资源路径
        example: ReadMe.md
        in: query
        name: path
        required: true
        type: string
      - description: Resource path Hash // 资源路径哈希
        example: hash123
        in: query
        name: pathHash
        required: true
        type: string
      - description: Vault name // 保险库名称
        example: defaultVault
        in: query
        name: vault
        required: true
        type: string
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.ShareCreateResponse'
              type: object
      security:
      - UserAuthToken: []
      summary: Query share by path
      tags:
      - Share
    post:
      consumes:
      - application/json
      description: Create a share token for a specific note or attachment, automatically
        resolve attachment references and authorize
      parameters:
      - description: Share Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.ShareCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.ShareCreateResponse'
              type: object
      security:
      - UserAuthToken: []
      summary: Create resource share
      tags:
      - Share
  /api/share/file:
    get:
      description: Get raw binary data of a specific attachment via share token
      parameters:
      - description: Auth Token
        in: header
        name: Share-Token
        required: true
        type: string
      - description: v3 entry UUID（与 id 二选一，优先） // v3 条目 UUID（与 id 二选一，优先）
        example: b3f1...uuid
        in: query
        name: entryId
        type: string
      - description: Resource ID (legacy int ID; v3 uses entryId) // 资源 ID（旧整型；v3
          用 entryId）
        example: 1
        in: query
        name: id
        type: integer
      - description: Share password // 分享密码
        example: "123456"
        in: query
        name: password
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Success
          schema:
            type: file
      security:
      - ShareAuthToken: []
      summary: Get shared attachment content
      tags:
      - Share
  /api/share/note:
    get:
      description: Get specific note content (restricted read-only access) via share
        token
      parameters:
      - description: Auth Token
        in: header
        name: Share-Token
        required: true
        type: string
      - description: v3 entry UUID（与 id 二选一，优先） // v3 条目 UUID（与 id 二选一，优先）
        example: b3f1...uuid
        in: query
        name: entryId
        type: string
      - description: Resource ID (legacy int ID; v3 uses entryId) // 资源 ID（旧整型；v3
          用 entryId）
        example: 1
        in: query
        name: id
        type: integer
      - description: Share password // 分享密码
        example: "123456"
        in: query
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteDTO'
              type: object
      security:
      - ShareAuthToken: []
      summary: Get shared note details
      tags:
      - Share
  /api/share/password:
    post:
      consumes:
      - application/json
      description: Set or update password and expiration time for a share record
      parameters:
      - description: Update Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.SharePasswordUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Update share password
      tags:
      - Share
  /api/share/short_link:
    post:
      consumes:
      - application/json
      description: Call sink.cool API to generate a short link for a given share record
      parameters:
      - description: Short Link Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.ShareShortLinkCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  type: string
              type: object
      security:
      - UserAuthToken: []
      summary: Create short link for share
      tags:
      - Share
  /api/shares:
    get:
      description: Get all active and inactive shares of the user, supports sorting
        and pagination
      parameters:
      - description: 'Sort field: created_at, updated_at, expires_at (default: created_at)'
        in: query
        name: sort_by
        type: string
      - description: 'Sort direction: asc or desc (default: desc)'
        in: query
        name: sort_order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.ShareListItem'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: List shares
      tags:
      - Share
...
securityDefinitions:
  ShareAuthToken:
    in: header
    name: Share-Token
    type: apiKey
  UserAuthToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"